[FEATURE] Value / Eel Expression / TypoScript-Object interoperability
Currently, there is a big distinction between simple values and Eel
expressions, and sub-TypoScript objects; despite them sharing the same
namespace.

However, basically, all of the above mentioned three TS data types produce
a value which they return to their parent.

This functionality must be tested in the following three dimensions:

1.a using Runtime::evaluate()
1.b using $this->tsValue() inside a TypoScript object

2.a without processors
2.b with processors

3.a pass in a simple type
3.b pass in an Eel Expression
3.c pass in a TypoScript object

This change implements the functionality in a test-driven way, testing
all the 12 cases from above.

This change also removes all references to "implementationClassName" which
was deprecated for a long time now as well. Using
"implementationClassName" instead of "@class" was broken anyway so this
is not considered breaking.

Resolves: #48359
Change-Id: I3f2569025808f7fa13bee5164b94cbfc982889bf
Reviewed-on: https://review.typo3.org/21003
Reviewed-by: Christian Mueller
Tested-by: Christian Mueller
Original-Commit-Hash: af27c1837f290746ff497656692d089807d0be24