[BUGFIX] Make sure prototype inheritance is only flattened at the end of all include-parsing
Problem Description
===================

Imagine a TypoScript file Base.ts2 with the following contents:

  prototype(A) {
    a = 'a'
  }
  prototype(B) < prototype(A)

… and another TypoScript file with:

  include: resource://…/Base.ts2
  prototype(A).a = 'new'

Then you would expect prototype(B).a to be "new" as well.

Without this change, this was not the case, because TypoScript
object prototypes were flattened too early.

Solution
========

Just make sure the flattening of the TypoScript prototypes
happens only at the very end of the full parsing step.

While I restructured the tests a bit more, they do not exactly
test this case, because it only broke for resource:// URLs
(which I don't know how to emulate in the tests), while it
worked for relative includes.

Change-Id: I3e1f408d8069a53df64b2a62181ed401dbe2f2fa
Reviewed-on: https://review.typo3.org/24493
Tested-by: Mr. Hudson
Tested-by: Sebastian Kurfuerst
Reviewed-by: Mattias Nilsson
Tested-by: Mattias Nilsson
Reviewed-by: Jacob Floyd
Reviewed-by: Sebastian Kurfuerst
Original-Commit-Hash: b908915322400022037808e485b39c68cc76d640