[BUGFIX] Support debug mode for CaseImplementation
The Case object did not match result correctly in debug mode, since the
rendered output of a matcher is annotated with debug comments and does
not equal the MATCH_NORESULT constant.

This change strips these comments from the rendered output if the debug
mode is enabled before comparing the strings.

To test the behavior a setter for the debug mode was introduced to the
TypoScript Runtime and an option was added to the TypoScriptView.

Change-Id: Ic0d24ddfb35a62a49f4e1d959a4dbe892c260d14
Fixes: #52923
Reviewed-on: https://review.typo3.org/24955
Tested-by: Mr. Hudson
Reviewed-by: Dominique Feyer
Tested-by: Dominique Feyer
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek
Reviewed-by: Sebastian Kurfuerst
Tested-by: Sebastian Kurfuerst
Original-Commit-Hash: 405dd1c3cb50c0dc46526cdc3f1b0b24388b0fd7