[BUGFIX] Default namespace is not available in included TS files
A default namespace (or any other namespace) defined in the TypoScript
parser is not defined for included TypoScript files because the parser
will create a whole new instance when parsing the include. This change
fixes this behaviour by creating a clone of the parser instead.

Releases: master, 1.0
Change-Id: Idaa3e0c5c3393533158840dc0617676f2589c628
Reviewed-on: https://review.typo3.org/26165
Reviewed-by: Robert Lemke
Tested-by: Robert Lemke
Original-Commit-Hash: 5aae72252300074f5d65b13d3806498e6d7be008