[!!!][TASK] Do not trim rendered strings
To make the output more predictable and allow for spaces at the
start and end of strings rendered in TypoScript trim was removed.

If you rely on your output to be trimmed this is breaking. You
can get the same behavior if you apply a String.trim() processor
to the TypoScript object for which the output should be trimmed.

Change-Id: I6c0654b63a78903d3097dec9fb20b3373bf8a400
Reviewed-on: https://review.typo3.org/25922
Reviewed-by: Christian Mueller
Tested-by: Christian Mueller
Reviewed-by: Mattias Nilsson
Tested-by: Mattias Nilsson
Tested-by: Mr. Hudson
Reviewed-by: Aske Ertmann
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek
Original-Commit-Hash: 7e020d4ceada321a1ac5abdc6b0e755484eb845b