[FEATURE] TypoScript content cache
A nested cache implementation for the TypoScript runtime.

Every TypoScript path that is an object can be marked as cacheable with
a new @cache meta property:

	root.myPath.@cache.mode = 'cached'

This will create a cache entry for everything rendered inside this path
on the first execution if the "enableContentCache" is set to TRUE
for the runtime through the setting (TRUE is default here) or view
option.

The cache entry identifier is auto-generated from the current context if
not specified explicitly with "entryIdentifier" beneath the @cache
property:

	root.myPath.@cache.entryIdentifier {
		node = ${node}
	}

The identifier should be configured to include the minimal amount of
context information that is needed to have distinct cache entries.

Cached paths can be nested with different entryIdentifier configuration
to allow for the re-use of cached parts in different contexts (e.g. a
menu or sidebar that is included on many pages in Neos).

A cache entry will be tagged with "Everything" if the "entryTags"
configuration is not given. Custom tags should be set depending on the
content to have a more intelligent flushing strategy (which depends on
the application):

	root.myPath.@cache.entryTags {
		1 = ${'Node_' + documentNode.identifier}
	}

For dynamic content that should always be evaluated (regardless of being
nested inside cached content) the caching can explicitly be disabled:

	root.myPath.dynamic.@cache.mode = 'uncached'

These parts will be rendered using the TypoScript runtime with all
serialized context values, unless the "storeContext" property is
configured. This property should be configured to include the minimal
amount of context variables that are necessary to render the uncached
path:

	root.myPath.dynamic.@cache {
		mode = 'uncached'

		context {
			1 = 'node'
		}
	}

The change depends on If7b615939a4216e6c3178c5a4f4a2b0e7a156041 in
TYPO3.Flow for property mapping from object to simple type.
To store Node context values for uncached segments change
I747da3e32f89d4a6bde3886584e391a66201441d is needed in TYPO3.TYPO3CR.
Change Ia83bd0dc6ba50150463b84d24f546d1da682a048 integrates the
content cache with Neos for automatic cache flushing by tag on changes
to nodes. The logging depends on TYPO3.Flow commit
14bacb4a8c38068ab56fb6ec23981e2d26f0bbd6.

Change-Id: Iae8f812b61cd1796f2363f97353270d146a13969
Releases: master
Reviewed-on: https://review.typo3.org/27432
Reviewed-by: Robert Lemke
Tested-by: Robert Lemke
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek
Original-Commit-Hash: 08d2bde8fda47dbf2a8f3596a504184214c41916