[!!!][TASK] Change handling of empty, false and null attribute values
This change updates the rendering of boolean and null attribute values.
A boolean value is rendered as an empty attribute or no attribute
at all. A null or empty string attribute value will be rendered as an
empty attribute (HTML5 syntax).

This will be breaking if code relied on the previous conversion of values
or the output is used for XML, where the new meta property ``@allowEmpty``
should be used to disable empty attributes.

Change-Id: I9dbddf140bcb336daa09fbfc72b7bb4de608f678
Resolves: NEOS-275
Releases: master
Reviewed-on: http://review.typo3.org/32436
Reviewed-by: Christopher Hlubek <hlubek@networkteam.com>
Tested-by: Christopher Hlubek <hlubek@networkteam.com>
Reviewed-by: Christian Mueller <christian@kitsunet.de>
Tested-by: Christian Mueller <christian@kitsunet.de>
Original-Commit-Hash: e3fda907b680b52c07f9a7bef048661ef327728b