[BUGFIX] Exception causes fatal error in TypoScriptPathProxy __toString
If an exception occurs during handling of a rendering exception a fatal
error is thrown due to __toString methods not being allowed to throw
exceptions. To prevent that we catch those exceptions and try to render
them with a context depenent exception handler instead. If that also
fails we fall back to a custom handling of the exception.

Change-Id: I3fa9282943c239d973d6d787656e6dbd3656c1fc
Releases: master, 1.2
Fixes: NEOS-898
Reviewed-on: http://review.typo3.org/32830
Reviewed-by: Aske Ertmann <aertmann@gmail.com>
Tested-by: Aske Ertmann <aertmann@gmail.com>
Reviewed-by: Karsten Dambekalns <karsten@typo3.org>
Tested-by: Karsten Dambekalns <karsten@typo3.org>
Original-Commit-Hash: b31bf5e1ac9783d9d2519c0b25b9aa7ee9bf1c79