[TASK] Use TypoScript object for cacheIdentifiers
With this change the ``@cache.entryIdentifier`` is rendered as a
TypoScript object. If none is configured (as is with all current
cache configurations) it is assumed to be of the newly introduced
type ``TYPO3.TypoScript:GlobalCacheIdentifiers``.
This new prototype includes the ``baseUri`` and ``format`` keys
by default but can be extended as any other prototype.
This ensures that cached links always match the request format
and baseUri of the current rendering.

The baseUri includes the domain, http protocol and port.

Change-Id: I3d172b0ebfe54afc37b6eaa8397e8c2355a7d7da
Releases: master
Resolves: NEOS-852
Reviewed-on: http://review.typo3.org/36210
Reviewed-by: Christian Müller <christian@kitsunet.de>
Tested-by: Christian Müller <christian@kitsunet.de>
Reviewed-by: Aske Ertmann <aertmann@gmail.com>
Reviewed-by: Karsten Dambekalns <karsten@typo3.org>
Tested-by: Karsten Dambekalns <karsten@typo3.org>
Reviewed-by: Christopher Hlubek <hlubek@networkteam.com>
Tested-by: Christopher Hlubek <hlubek@networkteam.com>
Original-Commit-Hash: ee6ec66b05f1ac735832771f47701c958eb80fc4