[TASK] Allow ignoreProperties on all Array implementations
The CaseImplementation allows for the special meta property
``@ignoreProperties`` to define keys that are not used in evaluation.
This functionality can be useful for other types of implementations
that iterate over the properties.
The functionality was extracted and to make sure it works on all
implementations of
``\TYPO3\TypoScript\TypoScriptObjects\AbstractArrayTypoScriptObject``
They have a new property "ignoreProperties" now that can be used in
own implementations.
Using "sortNestedTypoScriptKeys()" of the ArrayImplementation will
ignore those properties already.

Change-Id: Ib378020fc1a371ababa000dc96eace50bb4e3cfb
Releases: master
Reviewed-on: http://review.typo3.org/36663
Reviewed-by: Aske Ertmann <aertmann@gmail.com>
Reviewed-by: Christian Müller <christian.mueller@typo3.org>
Tested-by: Christian Müller <christian.mueller@typo3.org>
Original-Commit-Hash: 3dbffe3ce3a6b053ed6dbe8cfa5333b73cd947e8