[BUGFIX] Uncached segment use configured context on first render
Uncached segments need to define the rendering context they need, but
on first rendering they still had access to the full context which
made it harder to spot errors in the context configuration as they
only appeared on the second rendering.
Now the context is adjusted to the configuration even for the first
rendering.

Change-Id: Iace3a6faf9752028ac219b7346a2546bdf0423cb
Fixes: NEOS-1127
Releases: master, 1.2
Reviewed-on: http://review.typo3.org/38072
Reviewed-by: Christian Müller <christian@kitsunet.de>
Tested-by: Christian Müller <christian@kitsunet.de>
Reviewed-by: Sebastian Kurfuerst <sebastian@typo3.org>
Reviewed-by: Aske Ertmann <aertmann@gmail.com>
Reviewed-by: Karsten Dambekalns <karsten@typo3.org>
Tested-by: Karsten Dambekalns <karsten@typo3.org>
Original-Commit-Hash: 797589213561c71d6c320ac0d54329bce0c268fa