[BUGFIX] Fix parse errors with TypoScript comments
This commit fixes problems with parsing "comments in comments" in TypoScript

Example::

    # */ text meant as comment

Adds a unit test and a fixture that reproduces the parsing error.
Adds a unit test for the split pattern used to identify the "type of comment".
Refactors unit test for the TypoScript parser.

Change-Id: I2fe2c5a050f4dc6ea22cc4004c0c2b759aff4716
Resolves: NEOS-864
Releases: master, 1.2, 1.1
Reviewed-on: http://review.typo3.org/38217
Reviewed-by: Aske Ertmann <aertmann@gmail.com>
Tested-by: Aske Ertmann <aertmann@gmail.com>
Original-Commit-Hash: d290814939bee1adc9c8423020d0310889ca65a4