[BUGFIX] Set status after sucessfully rendering a TypoScript object
Due to the bugfix done in ``I8abefe9f8fb399a186d93737e2396dd55fc9c452``
a regression was introduced as the ``lastEvaluationStatus`` is now
considered in processors to allow conditions. But as the status
is not set in a consistent manner you can end up with a successfully
rendered TypoScript object and ``lastEvaluationStatus`` set to FALSE.

This is fixed by setting the status after successfully rendering a
TypoScript object.

Change-Id: Id05233b004d2936987008eaa21cc4b65780b5dd6
Releases: master, 2.0
Reviewed-on: http://review.typo3.org/40072
Reviewed-by: Christopher Hlubek <hlubek@networkteam.com>
Tested-by: Christopher Hlubek <hlubek@networkteam.com>
Original-Commit-Hash: 57e3cf6844300524feaf4e7465c04660ea92645a