TASK: Don't cache segments with mode `dynamic` and disabled `entryDiscriminator`
With this change the caching can be disabled by setting the `entryDiscriminator`
to `false` when using Content Cache mode `dynamic`.

Previously a cache entry was created anyways with the `entryDiscriminator` casted
to an (in this case empty) string.

Background:
The Content Cache mode `dynamic` was introduced in order to allow for more flexible
caching behaviors depending on the context.
But one important feature did not work yet: Being able to *disable* the cache
for certain requests.
With this change performance can be improved by caching the display of an interactive
element (i.e. cache Forms for GET requests)

Related: #1630